be more precise on which libs might have which

+++ internalize.rb
@@ -3,6 +3,7 @@ require 'rbconfig'
 
 Dir.chdir File.dirname(__FILE__) + '/lib' do
   for file in Dir['*'] do
+    next if file =~ /rubygems_plugin/ # don't want to copy that one over
     new_file = Config::CONFIG['sitelibdir'] + '/' + file
     FileUtils.rm_rf new_file if File.exist?(new_file)
     FileUtils.cp_r file, new_file
